moved discipline data into separate file to separate it from the parser
